In Progress; Loop through Classes in PersonalAvatars package
Changed the .gitignore to include the lib folder where the necessary .jar files are.

These .jar files are needed because we do not want to use Maven. This is the simple solution.

The settings.json file in the .vscode folder tells vscode to use these .jar files for the needed reflections import. This import is for searching out the classes in the PersonalAvatars package.