In Progress; Loop through Classes in PersonalAvatars package
Added a user input for the default perception range.

Changed the perception in room.java getAdjacentToAvatar to be dynamic.

Added perceptionRange to the SuperAvatar Class. Needs to be given if a new Avatar is created.

Changed the name of TestAvatar to TemplateAvatar. Here we can provide a simple randomly moving avatar that others can use as a template for their own avatars.